#268 tidy-up handling of LMA ops and break-out tests to separate file.
